backend/requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/ML
openai>=1.56.0

# Vector Store
faiss-cpu==1.8.0
numpy==1.26.4

# Pinecone (optional)
pinecone-client==5.0.1

# Document processing
PyMuPDF==1.24.10
python-docx==1.1.2

# Caching
redis==5.1.1

# Validation
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# Utilities
python-dotenv==1.0.1
